rl_requirement.txt
accelerate
numpy==2.1.2
torchvision>=0.13.1
scipy
tqdm
huggingface_hub==0.24.7
Torch_Pruning
gymnasium==0.29.1
matplotlib==3.9.2
pandas==2.2.3
Pillow==10.4.0
stable_baselines3==2.3.2
six==1.16.0
wandb==0.18.3
tensorboard==2.18.0
scikit-image
diffusers


pip install scikit-image gymnasium==0.29.1 Pillow==10.4.0 stable_baselines3==2.3.2 six==1.16.0 